refactor: NotLoginHeader 공용컴포넌트로 분리. 관련 코드 수정
